Forum home controller refactoring

+++ Forum/Controllers/ForumHomeController.cs
@@ -1,15 +1,7 @@
 ﻿using AutoMapper;
-using Azure.Core;
-using Entities.DTO.ForumDto.Create;
-using Entities.DTO.ForumDto.ForumView;
 using Entities.ViewModels.Forum;
 using Interfaces.Forum;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-using Newtonsoft.Json;
-using Services;
-using System.Diagnostics;
-using System.Net;
 using System.Security.Claims;
 
 namespace Forum.Controllers
@@ -29,22 +21,25 @@ namespace Forum.Controllers
         {
             var model = await _forumService.GetForumCategoriesAndForumBasesForModel();
 
+            if(!User.Identity.IsAuthenticated)
+            {
+                return Unauthorized("Unauthorized access.");         
+            }
+
             return View("~/Views/Forum/ForumHome.cshtml", model);
         }
         public async Task<IActionResult> RedirectToCreateCategory()
         {
             return View("~/Views/Forum/Add/ForumAddCategory.cshtml");
         }
-        
-        /*public async Task<IActionResult> CreateForumCategory(ForumHomeViewModel model)
-        {
-            var user = _mapper.Map<ForumHomeViewModel>(model-);
-            return View("~/Views/Forum/Add/ForumAddForumBase.cshtml", modelD);
-        }*/
-        
         [Route("categories/{categoryId}/forums/{forumId}/topics", Name = "ForumTopics")]
         public async Task<IActionResult> ForumTopics(int categoryId, int forumId)
         {
+            if (!User.Identity.IsAuthenticated)
+            {
+                return Unauthorized("Unauthorized access.");
+            }
+
             var model = await _forumService.GetForumTopicsForModel(categoryId, forumId);
             await _forumService.IncreaseViewCounterForForumBase(categoryId, forumId);
 
@@ -55,6 +50,12 @@ namespace Forum.Controllers
         public async Task<IActionResult> TopicPosts(int categoryId, int forumId, int topicId, int pageId = 0)
         {
             int maxiumPostsPerPage = 4;
+
+            if (!User.Identity.IsAuthenticated)
+            {
+                return Unauthorized("Unauthorized access.");
+            }
+
             var model = await _forumService.GetTopicPostsForModel(categoryId, forumId, topicId, pageId, maxiumPostsPerPage);
 
             var tasks = model.Posts.Select(
@@ -77,6 +78,11 @@ namespace Forum.Controllers
         }
         public async Task<ActionResult> DeletePost(int categoryId, int forumId, int topicId, int postId, ForumTopicViewModel model)
         {
+            if (!User.Identity.IsAuthenticated)
+            {
+                return Unauthorized("Unauthorized access.");
+            }
+
             var res = await _forumService.DeleteForumPost(categoryId, forumId, topicId, postId);
             int totalPosts = 0;
 
@@ -99,6 +105,11 @@ namespace Forum.Controllers
         }
         public async Task<ActionResult> UpdatePost(int categoryId, int forumId, int topicId, int postId, int pageId, string newText)
         {
+            if (!User.Identity.IsAuthenticated)
+            {
+                return Unauthorized("Unauthorized access.");
+            }
+
             var res = await _forumService.UpdatePost(categoryId, forumId, topicId, postId, newText);
 
             return Json(new { redirectToUrl = Url.Action("TopicPosts", "ForumHome", new { categoryId = categoryId, forumId = forumId, topicId = topicId, pageId = pageId }) });
